gramian: the metric moves above the algebra it acts on
The exterior derivative, the wedge, the contraction, the transfer and
both pairings involve no metric; the inner product, the Hodge star and
the musical isomorphisms involve nothing else. That is invariant 5, and
it was documentation. It is now the boundary between two crates.

`multialgebra` drops its dependency on `gramian` and becomes exactly the
metric-free algebra. `Factor::gramian` becomes `induced_form`, taking a
bare bilinear form and returning one: an induced *form* needs only a
form, and non-degeneracy and signature are properties the algebra never
asks about. `gramian` gains the dependency and, with it, everything that
needs a metric: `variance_gramian`, `slot_gramian`, `tensor_gramian`, the
multi/multivector/multiform Gramians, the free `inner`, and a `TensorExt`
carrying `norm`, `hodge_star`, `star` and `musical`.

`inner` is free where the star is a method, for the reason the pairings
already are: an inner product is symmetric in its arguments and
privileges neither, while a star is a unary operator written postfix.

The laws move with the operations. `multialgebra`'s suite keeps the
metric-free ones and states them against plain matrices; the seven that
need a metric are `gramian`'s now. The counts add back to what they were,
which is the check that matters here: 457 tests before and after.

`Tensor::apply_to_slot` becomes public, being what the star and the
musicals need from one crate up, and `slots_mut` with it.
