+++ crates/derham/src/interpolate/form.rs
@@ -195,7 +195,8 @@ mod test {
   use super::*;
   use approx::assert_relative_eq;
   use gramian::{Gramian, Metric};
-  use multialgebra::{exterior_bases, exterior_dim, multiform_gramian};
+  use gramian::tensor::multiform_gramian;
+  use multialgebra::{exterior_bases, exterior_dim};
   use multiindex::combinations;
   use simplicial::atlas::{SimplexQuadRule, unit_simplex_volume};
   use simplicial::linalg::Vector;
