+++ crates/multialgebra/Cargo.toml
@@ -11,7 +11,6 @@ keywords = ["exterior-algebra", "symmetric-algebra", "multilinear", "tensor"]
 categories = ["mathematics"]
 
 [dependencies]
-gramian = { path = "../gramian", version = "0.2.0" }
 multiindex = { path = "../multiindex", version = "0.2.0" }
 
 nalgebra = "0.35.0"
